adding delete by id not found logic

+++ src/main/java/guru/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -61,9 +61,12 @@ public class BeerServiceJPA implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID beerId) {
-
-        beerRepository.deleteById(beerId);
+    public Boolean deleteById(UUID beerId) {
+        if (beerRepository.existsById(beerId)) {
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
